refactor(ms5611): compensate in decoder
